Fixed no value assinged to inputDir issue
After targeting the latest version 0.0.13, I'm getting a no value assigned to 'inputDir' error.

Possible fix for this issue is to assign a value to 'inputDir' or mark the field as optional. But I can see that this field is not used anywhere in this file.
Is this fix okay? If yes then please push this changes to 0.0.13